Moved serving endpoints services to a new controller and added JavaDoc
