+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/ET1ServingController.java
@@ -1,0 +1,109 @@
+package uk.gov.hmcts.ethos.replacement.docmosis.controllers;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.et.common.model.ccd.CCDCallbackResponse;
+import uk.gov.hmcts.et.common.model.ccd.CCDRequest;
+import uk.gov.hmcts.et.common.model.ccd.CaseData;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.ServingHelper;
+import uk.gov.hmcts.ethos.replacement.docmosis.service.VerifyTokenService;
+
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static uk.gov.hmcts.ethos.replacement.docmosis.helpers.CallbackRespHelper.getCallbackRespEntityNoErrors;
+
+
+@Slf4j
+@RequiredArgsConstructor
+@RestController
+public class ET1ServingController {
+
+    private static final String INVALID_TOKEN = "Invalid Token {}";
+
+    private final VerifyTokenService verifyTokenService;
+
+    /**
+     * This service Gets userToken as a parameter for security validation
+     * and ccdRequest data which has caseData as an object.
+     * It is used to populate
+     * @param  userToken        Used for authorisation
+     *
+     * @param ccdRequest        CaseData which is a generic data type for most of the
+     *                          methods which holds ET1 case data
+     * @return ResponseEntity   It is an HTTPEntity response which has CCDCallbackResponse that
+     *                          includes caseData which contains the upload document names of
+     *                          type "Another type of document" in a html string format.
+     */
+    @PostMapping(value = "/midServingDocumentOtherTypeNames", consumes = APPLICATION_JSON_VALUE)
+    @Operation(summary = "return serving document other type names")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Accessed successfully",
+                    content = {
+                            @Content(mediaType = "application/json", schema = @Schema(implementation = CCDCallbackResponse.class))
+                    }),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "500", description = "Internal Server Error")
+    })
+    public ResponseEntity<CCDCallbackResponse> midServingDocumentOtherTypeNames(
+            @RequestBody CCDRequest ccdRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
+
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error(INVALID_TOKEN, userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
+        CaseData caseData = ccdRequest.getCaseDetails().getCaseData();
+        caseData.setOtherTypeDocumentName(ServingHelper.generateOtherTypeDocumentName(caseData.getServingDocumentCollection()));
+
+        return getCallbackRespEntityNoErrors(ccdRequest.getCaseDetails().getCaseData());
+    }
+
+    /**
+     * This service Gets userToken as a parameter for security validation
+     * and ccdRequest data which has caseData as an object.
+     * It is used to populate
+     * @param  userToken        Used for authorisation
+     *
+     * @param ccdRequest        CaseData which is a generic data type for most of the
+     *                          methods which holds ET1 case data
+     * @return ResponseEntity   It is an HTTPEntity response which has CCDCallbackResponse that
+     *                          includes caseData which contains the claimant and respondent
+     *                          addresses in html string format, and an email link that will
+     *                          be used to send email to Acas in MailTo syntax.
+     */
+    @PostMapping(value = "/midServingDocumentRecipient", consumes = APPLICATION_JSON_VALUE)
+    @Operation(summary = "return serving document other type recipient's addresses")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Accessed successfully",
+                    content = {
+                            @Content(mediaType = "application/json", schema = @Schema(implementation = CCDCallbackResponse.class))
+                    }),
+            @ApiResponse(responseCode = "400", description = "Bad Request"),
+            @ApiResponse(responseCode = "500", description = "Internal Server Error")
+    })
+    public ResponseEntity<CCDCallbackResponse> midServingDocumentRecipient(
+            @RequestBody CCDRequest ccdRequest,
+            @RequestHeader(value = "Authorization") String userToken) {
+        if (!verifyTokenService.verifyTokenSignature(userToken)) {
+            log.error(INVALID_TOKEN, userToken);
+            return ResponseEntity.status(FORBIDDEN.value()).build();
+        }
+
+        CaseData caseData = ccdRequest.getCaseDetails().getCaseData();
+        caseData.setClaimantAndRespondentAddresses(ServingHelper.generateClaimantAndRespondentAddress(caseData));
+        caseData.setEmailLinkToAcas(ServingHelper.generateEmailLinkToAcas(caseData));
+
+        return getCallbackRespEntityNoErrors(caseData);
+    }
+}
