+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/CaseActionsForCaseWorkerControllerTest.java
@@ -113,8 +113,7 @@ public class CaseActionsForCaseWorkerControllerTest {
     private static final String DYNAMIC_JUDGMENT_URL = "/dynamicJudgments";
     private static final String JUDGEMENT_SUBMITTED_URL = "/judgementSubmitted";
     private static final String REINSTATE_CLOSED_CASE_MID_EVENT_VALIDATION_URL = "/reinstateClosedCaseMidEventValidation";
-    private static final String SERVING_DOCUMENT_OTHER_TYPE_NAMES_URL = "/midServingDocumentOtherTypeNames";
-    private static final String SERVING_DOCUMENT_RECIPIENT_URL = "/midServingDocumentRecipient";
+
 
     @Autowired
     private WebApplicationContext applicationContext;
@@ -177,7 +176,6 @@ public class CaseActionsForCaseWorkerControllerTest {
     private JsonNode requestContent;
     private JsonNode requestContent2;
     private JsonNode requestContent3;
-    private JsonNode requestContent4;
     private SubmitEvent submitEvent;
     private DefaultValues defaultValues;
 
@@ -189,8 +187,6 @@ public class CaseActionsForCaseWorkerControllerTest {
                 .getResource("/exampleV2.json")).toURI()));
         requestContent3 = objectMapper.readTree(new File(Objects.requireNonNull(getClass()
                 .getResource("/exampleV3.json")).toURI()));
-        requestContent4 = objectMapper.readTree(new File(Objects.requireNonNull(getClass()
-                .getResource("/exampleV4.json")).toURI()));
 
         objectMapper.readTree(new File(Objects.requireNonNull(getClass()
                 .getResource("/CaseCloseEvent_ValidHearingStatusCaseDetails.json")).toURI()));
@@ -797,50 +793,6 @@ public class CaseActionsForCaseWorkerControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data", notNullValue()))
                 .andExpect(jsonPath("$.errors", hasSize(1)))
-                .andExpect(jsonPath("$.warnings", nullValue()));
-    }
-
-    @Test
-    public void midServingDocumentOtherTypeNames() throws Exception {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-
-        String expectedDocumentName = "**<big>test-filename.xlsx</big>**<br/><small>Test description</small><br/>";
-        mvc.perform(post(SERVING_DOCUMENT_OTHER_TYPE_NAMES_URL)
-                        .content(requestContent4.toString())
-                        .header("Authorization", AUTH_TOKEN)
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.data.otherTypeDocumentName").value(expectedDocumentName))
-                .andExpect(jsonPath("$.errors", nullValue()))
-                .andExpect(jsonPath("$.warnings", nullValue()));
-    }
-
-    @Test
-    public void midServingDocumentRecipient() throws Exception {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(true);
-
-        String expectedClaimantAndRespondentAddress = "**<big>Claimant</big>**<br/>Doris Johnson"
-                + "<br/>232 Petticoat Square<br/>22 House<br/>London<br/>W10 4AG<br/><br/>"
-                + "**<big>Respondent 1</big>**<br/>Antonio Vazquez"
-                + "<br/>11 Small Street<br/>22 House<br/>Manchester<br/>M12 42R<br/><br/>"
-                + "**<big>Respondent 2</big>**<br/>Juan Garcia<br/>12 Small Street<br/>24 House"
-                + "<br/>Manchester<br/>M12 4ED<br/><br/>";
-
-        String expectedLink = "mailto:ET3@acas.org.uk?subject=2120001/2019" +
-                "&body=Parties%20in%20claim%3A%20Doris%20Johnson%20vs%20Antonio%20Vazquez%2C%20Juan%20Garcia%0D%0A" +
-                "Case%20reference%20number%3A%202120001/2019%0D%0A%0D%0ADear%20Acas%2C%0D%0A%0D%0AThe%20tribunal%20" +
-                "has%20completed%20ET1%20serving%20to%20the%20respondent.%0D%0A%0D%0AThe%20documents%20we%20sent%20are" +
-                "%20attached%20to%20this%20email.%0D%0A%0D%0A";
-
-        mvc.perform(post(SERVING_DOCUMENT_RECIPIENT_URL)
-                        .content(requestContent4.toString())
-                        .header("Authorization", AUTH_TOKEN)
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.data.claimantAndRespondentAddresses")
-                        .value(expectedClaimantAndRespondentAddress))
-                .andExpect(jsonPath("$.data.emailLinkToAcas").value(expectedLink))
-                .andExpect(jsonPath("$.errors", nullValue()))
                 .andExpect(jsonPath("$.warnings", nullValue()));
     }
 
@@ -1662,25 +1614,5 @@ public class CaseActionsForCaseWorkerControllerTest {
 
         verify(clerkService, never()).initialiseClerkResponsible(isA(CaseData.class));
         verify(fileLocationSelectionService, never()).initialiseFileLocation(isA(CaseData.class));
-    }
-
-    @Test
-    public void midServingDocumentOtherTypeNamesForbidden() throws Exception {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(false);
-        mvc.perform(post(SERVING_DOCUMENT_OTHER_TYPE_NAMES_URL)
-                        .content(requestContent4.toString())
-                        .header("Authorization", AUTH_TOKEN)
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isForbidden());
-    }
-
-    @Test
-    public void midServingDocumentRecipientForbidden() throws Exception {
-        when(verifyTokenService.verifyTokenSignature(AUTH_TOKEN)).thenReturn(false);
-        mvc.perform(post(SERVING_DOCUMENT_RECIPIENT_URL)
-                        .content(requestContent4.toString())
-                        .header("Authorization", AUTH_TOKEN)
-                        .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isForbidden());
     }
 }
